2.	Realizar un programa que convierta una cantidad de metros dada en su equivalente en Pies, Yardas y Millas.

+++ src/pp2/problema2/PP2Problema2.java
@@ -62,7 +62,7 @@ public class PP2Problema2 {
         double miDato;
         Scanner entradaEscaner = new Scanner(System.in);
         System.out.print("Introduce el dato " + d + "\n");
-        miDato= entradaEscaner.nextInt();
+        miDato= entradaEscaner.nextDouble();
         return miDato;
     }
     public static void muestraResultado(double a){
